+++ lib/pmdtester/builders/pmd_report_builder.rb
@@ -1,5 +1,9 @@
 require 'fileutils'
 require_relative '../cmd'
+require_relative '../project'
+require_relative '../pmd_branch_detail'
+require_relative '../pmd_report_detail'
+
 include PmdTester
 module PmdTester
   # Building pmd xml reports according to a list of standard
@@ -11,11 +15,9 @@ module PmdTester
       @local_git_repo = local_git_repo
       @pmd_branch_name = pmd_branch_name
       @pwd = Dir.getwd
-    end
 
-    def create_repositories_dir
-      @repositories_dir = "#{@pwd}/target/repositories"
-      FileUtils.mkdir_p(@repositories_dir) unless File.directory?(@repositories_dir)
+      @pmd_branch_details = PmdBranchDetail.new
+      @pmd_branch_details.branch_name = pmd_branch_name
     end
 
     def execute_reset_cmd(type, tag)
@@ -34,17 +36,14 @@ module PmdTester
     def get_projects
       puts 'Cloning projects started'
 
-      create_repositories_dir
-
       @projects.each do |project|
-        path = "#{@repositories_dir}/#{project.name}"
+        path = project.local_source_path
         clone_cmd = "#{project.type} clone #{project.connection} #{path}"
         if File.exist?(path)
           puts "Skipping clone, project path #{path} already exists"
         else
           Cmd.execute(clone_cmd)
         end
-        project.local_path = path
 
         next if project.tag.nil?
         Dir.chdir(path) do
@@ -58,6 +57,9 @@ module PmdTester
         checkout_cmd = "git checkout #{@pmd_branch_name}"
         Cmd.execute(checkout_cmd)
 
+        @pmd_branch_details.branch_last_sha = get_last_commit_sha
+        @pmd_branch_details.branch_last_message = get_last_commit_message
+
         package_cmd = './mvnw clean package -Dpmd.skip=true -Dmaven.test.skip=true' \
                       ' -Dmaven.checkstyle.skip=true -Dmaven.javadoc.skip=true'
         Cmd.execute(package_cmd)
@@ -66,34 +68,50 @@ module PmdTester
                       '--non-recursive org.codehaus.mojo:exec-maven-plugin:1.5.0:exec'
         @pmd_version = Cmd.execute(version_cmd)
 
-        target_dir = "#{@pwd}/target"
-        unzip_cmd = "unzip -qo pmd-dist/target/pmd-bin-#{@pmd_version}.zip -d #{target_dir}"
+        unzip_cmd = "unzip -qo pmd-dist/target/pmd-bin-#{@pmd_version}.zip -d #{@pwd}/target"
         Cmd.execute(unzip_cmd)
       end
+    end
+
+    def get_last_commit_sha
+      get_last_commit_sha_cmd = 'git rev-parse HEAD'
+      Cmd.execute(get_last_commit_sha_cmd)
+    end
+
+    def get_last_commit_message
+      get_last_commit_message_cmd = 'git log -1 --pretty=%B'
+      Cmd.execute(get_last_commit_message_cmd)
     end
 
     def generate_pmd_report(src_root_dir, report_file)
       run_path = "target/pmd-bin-#{@pmd_version}/bin/run.sh"
       pmd_cmd = "#{run_path} pmd -d #{src_root_dir} -f xml -R #{@branch_config} " \
                 "-r #{report_file} -failOnViolation false"
+      start_time = Time.now
       Cmd.execute(pmd_cmd)
+      end_time = Time.now
+      [end_time - start_time, end_time]
     end
 
     def generate_pmd_reports
       puts "Generating pmd Report started -- branch #{@pmd_branch_name}"
-
       get_pmd_binary_file
 
-      pmd_branch_name = @pmd_branch_name.delete('/')
-      branch_file = "target/reports/#{pmd_branch_name}"
-      FileUtils.mkdir_p(branch_file) unless File.directory?(branch_file)
-
+      sum_time = 0
       @projects.each do |project|
-        project_report_file = "#{branch_file}/#{project.name}.xml"
-        project_source_dir = "target/repositories/#{project.name}"
-        generate_pmd_report(project_source_dir, project_report_file)
-        project.pmd_reports.store(@pmd_branch_name, project_report_file)
+        execution_time, end_time =
+          generate_pmd_report(project.local_source_path,
+                              project.get_pmd_report_path(@pmd_branch_name))
+        sum_time += execution_time
+
+        report_details = PmdReportDetail.new
+        report_details.execution_time = execution_time
+        report_details.timestamp = end_time
+        report_details.save(project.get_report_info_path(@pmd_branch_name))
       end
+
+      @pmd_branch_details.execution_time = sum_time
+      @pmd_branch_details.save
     end
 
     def build