+++ test/integration_test_runner.rb
@@ -12,10 +12,10 @@ class IntegrationTestRunner < Test::Unit::TestCase
     `bundle exec bin/pmdtester #{argv}`
 
     assert_equal(0, $CHILD_STATUS.exitstatus)
-    assert_path_exist('target/reports/master/checkstyle.xml')
-    assert_path_exist('target/reports/master/pmd.xml')
-    assert_path_exist('target/reports/pmd_releases6.1.0/checkstyle.xml')
-    assert_path_exist('target/reports/pmd_releases6.1.0/pmd.xml')
+    assert_path_exist('target/reports/master/checkstyle/pmd_report.xml')
+    assert_path_exist('target/reports/master/pmd/pmd_report.xml')
+    assert_path_exist('target/reports/pmd_releases/6.1.0/checkstyle/pmd_report.xml')
+    assert_path_exist('target/reports/pmd_releases/6.1.0/pmd/pmd_report.xml')
     assert_path_exist('target/reports/diff/checkstyle/index.html')
     assert_path_exist('target/reports/diff/pmd/index.html')
   end
@@ -27,8 +27,8 @@ class IntegrationTestRunner < Test::Unit::TestCase
     `bundle exec bin/pmdtester #{argv}`
 
     assert_equal(0, $CHILD_STATUS.exitstatus)
-    assert_path_exist('target/reports/pmd_releases6.1.0/checkstyle.xml')
-    assert_path_exist('target/reports/pmd_releases6.1.0/pmd.xml')
+    assert_path_exist('target/reports/pmd_releases/6.1.0/checkstyle/pmd_report.xml')
+    assert_path_exist('target/reports/pmd_releases/6.1.0/pmd/pmd_report.xml')
     assert_path_exist('target/reports/diff/checkstyle/index.html')
     assert_path_exist('target/reports/diff/pmd/index.html')
   end