+++ test/test_diff_builder.rb
@@ -1,17 +1,23 @@
 require 'test/unit'
 require_relative '../lib/pmdtester/builders/diff_builder'
+require_relative '../lib/pmdtester/builders/html_report_builder'
 
 # Unit test class for PmdTester::DiffBuilder
 class TestDiffBuilder < Test::Unit::TestCase
+  include PmdTester
+  BASE_REPORT_INFO_PATH = 'test/resources/html_report_builder/base_report_info.json'.freeze
+  PATCH_REPORT_INFO_PATH = 'test/resources/html_report_builder/patch_report_info.json'.freeze
+
   def setup
     `rake clean`
   end
 
   def test_violation_diffs
-    diff_builder = PmdTester::DiffBuilder.new
-    base_report_path = 'test/resources/test_violation_diffs_base.xml'
-    patch_report_path = 'test/resources/test_violation_diffs_patch.xml'
-    diffs_report = diff_builder.build(base_report_path, patch_report_path)
+    diff_builder = DiffBuilder.new
+    base_report_path = 'test/resources/diff_builder/test_violation_diffs_base.xml'
+    patch_report_path = 'test/resources/diff_builder/test_violation_diffs_patch.xml'
+    diffs_report = diff_builder.build(base_report_path, patch_report_path,
+                                      BASE_REPORT_INFO_PATH, PATCH_REPORT_INFO_PATH)
     violation_diffs = diffs_report.violation_diffs
     error_diffs = diffs_report.error_diffs
     keys = violation_diffs.keys
@@ -26,13 +32,16 @@ class TestDiffBuilder < Test::Unit::TestCase
     assert_equal('Patch1.java', keys[2])
     assert_equal('Patch2.java', keys[3])
     assert_equal('Patch3.java', keys[4])
+
+    assert_equal('00:00:56', diffs_report.diff_execution_time)
   end
 
   def test_error_diffs
-    diff_builder = PmdTester::DiffBuilder.new
-    base_report_path = 'test/resources/test_error_diffs_base.xml'
-    patch_report_path = 'test/resources/test_error_diffs_patch.xml'
-    diffs_report = diff_builder.build(base_report_path, patch_report_path)
+    diff_builder = DiffBuilder.new
+    base_report_path = 'test/resources/diff_builder/test_error_diffs_base.xml'
+    patch_report_path = 'test/resources/diff_builder/test_error_diffs_patch.xml'
+    diffs_report = diff_builder.build(base_report_path, patch_report_path,
+                                      BASE_REPORT_INFO_PATH, PATCH_REPORT_INFO_PATH)
     violation_diffs = diffs_report.violation_diffs
     error_diffs = diffs_report.error_diffs
     keys = error_diffs.keys